[319. Bulb Switcher][Accepted]committed by LeetCode Extension

+++ 319-Bulb-Switcher/solution.java
@@ -6,7 +6,7 @@ public class Solution {
          *  Else it has even number of factors, i.e., 7 has 1 and 7, and 8 has 1,2,4,8, then the ith bulb is off after n round.
          */
         // just count the square numbers
-        return Math.sqrt(n);
+        return (int)Math.sqrt(n);
         /*if (n == 0) return 0;
         int i;
         for (i = 1; i * i <= n; i++) {
